requirements.txt
colorama
requests
dnspython
python-whois
